+++ src/components/Button/Button.type.ts
@@ -28,6 +28,7 @@ export type ButtonType = {
   children?: React.ReactNode,
   className?: string,
   href?: string,
+  type?: 'button' | 'submit' | 'reset',
   style?: ButtonStyles,
   padding?: string,
   gap?: string,