+++ src/components/NavbarLeft/NavbarLeft.type.ts
@@ -1,0 +1,8 @@
+/**
+ * NavbarLeftType type for NavbarLeft component
+ * @example
+ * <NavbarLeft toggleDialog={() => {}} />
+ */
+export type NavbarLeftType = {
+  toggleDialog?: () => void;
+};